Put /me messages into the backlog of messages.
In poke.py, note that dictionaries in each channels' backlog gained
a new field, 'type', which can be set with the optional parameter 'kind'
of #log_to_backlog (defaults to 'msg').

In script.js, addLine gained the ability to pass an array of classes.
Passing a single string still works but should be used to pass a single
class.


Former-commit-id: 865b4228388cdc5770054aacf6bc256598a2c05a